Second exercise

+++ main.py
@@ -4,6 +4,7 @@ import sys
 import requests
 from PyQt6.QtGui import QPixmap
 from PyQt6.QtWidgets import QApplication, QWidget, QLabel
+from PyQt6.QtCore import Qt
 
 SCREEN_SIZE = [600, 450]
 
@@ -14,6 +15,8 @@ class Example(QWidget):
         self.longitude = -64.825251
         self.lattitude = 18.299605
         self.z = 16
+        self.z_min = 0
+        self.z_max = 21
         self.getImage()
         self.initUI()
 
@@ -34,7 +37,6 @@ class Example(QWidget):
             print("Http статус:", response.status_code, "(", response.reason, ")")
             sys.exit(1)
 
-        # Запишем полученное изображение в файл.
         self.map_file = "map.png"
         with open(self.map_file, "wb") as file:
             file.write(response.content)
@@ -50,10 +52,31 @@ class Example(QWidget):
         self.image.resize(600, 450)
         self.image.setPixmap(self.pixmap)
 
+    def updateMap(self):
+        self.pixmap = QPixmap(self.map_file)
+        self.image.setPixmap(self.pixmap)
+
+
+    def keyPressEvent(self, event):
+        key = event.key()
+        if key == Qt.Key.Key_PageUp:
+            self.z += 1
+            if self.z > self.z_max:
+                self.z = self.z_max
+            self.getImage()
+            self.updateMap()
+
+        if key == Qt.Key.Key_PageDown:
+            self.z -= 1
+            if self.z < self.z_min:
+                self.z = self.z_min
+            self.getImage()
+            self.updateMap()
+
     def closeEvent(self, event):
         """При закрытии формы подчищаем за собой"""
-        os.remove(self.map_file)
-
+        if os.path.exists(self.map_file):
+            os.remove(self.map_file)
 
 if __name__ == '__main__':
     app = QApplication(sys.argv)
